requirements.txt
# Core Agent Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

# Turkish Financial Data
borsapy>=0.1.0
evdspy>=1.1.0

# RAG Components
lancedb>=0.15.0
sentence-transformers>=3.0.0

# Evaluation
ragas>=0.2.0
deepeval>=1.0.0

# Experiment Tracking
wandb>=0.18.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0

# UI
streamlit>=1.30.0
